+++ chat-web-client/src/hooks/index.ts
@@ -16,3 +16,4 @@ export * from './useNotifications';
 export * from './useAI';
 export * from './useSubscription';
 export * from './useRefreshProfile';
+export * from './usePushNotifications';